Added RepeatedField.
Unfortunately, this causes some problems with the CPP code generator, which does not seem to be able to handle generic classes that call methods that should be made available by the type constraints. A temporary fix for this has been implemented by making the build script overwrite the offending output C++ file.